refactor: extract model info lookup
